IO multiplexer

+++ srcs/Server/Server.cpp
@@ -23,7 +23,7 @@
 
 namespace {
 
-const int MAX_SESSION = 128;
+const int MAX_SESSION = 2;
 
 
 // void stop_by_signal(int sig) {
@@ -54,7 +54,6 @@ const int MAX_SESSION = 128;
 // ServerResult set_signal() {
 //     // return ServerResult::ok(OK);
 //
-//     // todo
 // 	errno = 0;
 // 	if (signal(SIGABRT, stop_by_signal) == SIG_ERR) {
 // 		const std::string error_msg = CREATE_ERROR_INFO_ERRNO(errno);
@@ -251,9 +250,12 @@ Result<IOMultiplexer *, std::string> Server::create_io_multiplexer_fds() {
 
 ServerResult Server::run() {
 	while (true) {
+        // char *p = new char[100]; (void)p;
+        // sleep(1);
         DEBUG_SERVER_PRINT(" run 1 timeout management");
         management_timeout_events();
         set_io_timeout();
+
         DEBUG_SERVER_PRINT(" run 2 get_io_ready_fd");
         ServerResult fd_ready_result = this->fds_->get_io_ready_fd();
         DEBUG_SERVER_PRINT(" run 3 ready result");
@@ -265,7 +267,6 @@ ServerResult Server::run() {
 		int ready_fd = fd_ready_result.ok_value();
         DEBUG_SERVER_PRINT(" run 4 ready_fd: %d", ready_fd);
 		if (ready_fd == IO_TIMEOUT) {
-            // std::cerr << "[Server INFO] timeout" << std::endl;
             if (this->echo_mode_on_) {
                 DEBUG_SERVER_PRINT("  timeout -> break");
                 break;
@@ -300,38 +301,79 @@ void Server::erase_from_timeout_manager(int cgi_fd) {
 }
 
 
-// todo: timeout read/write fd
-// read kill -> write ?
-// write kill ->
-void Server::management_timeout_events() {
-    // cgi event
-    time_t current_time = std::time(NULL);
-    DEBUG_PRINT(GREEN, "management_timeout current: %zu", current_time);
-    std::set<FdTimeoutLimitPair>::const_iterator cgi;
-
-    DEBUG_PRINT(GREEN, " debug print cgi_sessions:[");
+void Server::management_cgi_executing_timeout(time_t current_time) {
+    std::ostringstream cgi_sessions;
+    cgi_sessions << "debug print cgi_sessions:[";
     for (std::map<CgiFd, Event *>::iterator itr = cgi_events_.begin(); itr != cgi_events_.end(); ++itr) {
-        DEBUG_PRINT(GREEN, " fd:%d, client:%p", itr->first, itr->second);
+        cgi_sessions << "fd:" << itr->first << ", client:" << itr->second << " ";
     }
-    DEBUG_PRINT(GREEN, "]");
+    cgi_sessions << "]";
+    DEBUG_PRINT(GREEN, "%s", cgi_sessions.str().c_str());
 
+    std::ostringstream cgi_fds;
+    cgi_fds << "debug print cgi_fds:[";
+    for (std::set<FdTimeoutLimitPair>::iterator itr = this->cgi_fds_.begin(); itr != this->cgi_fds_.end(); ++itr) {
+        cgi_fds << itr->second << " ";
+    }
+    cgi_fds << "]";
+    DEBUG_PRINT(GREEN, "%s", cgi_fds.str().c_str());
+
+    std::set<FdTimeoutLimitPair>::const_iterator cgi;
     for (cgi = this->cgi_fds_.begin(); cgi != this->cgi_fds_.end(); ++cgi) {
         time_t timeout_limit = cgi->first;
         DEBUG_SERVER_PRINT(" cgi_fd: %d, time limit: %zu, current: %zu -> %s",
                            cgi->second, timeout_limit, current_time, (timeout_limit <= current_time ? "limited" : "ok"));
         if (current_time < timeout_limit) {
-            break;
+            break;  // sorted
         }
 
         int cgi_fd = cgi->second;
         Event *client = this->cgi_events_[cgi_fd];
-        DEBUG_PRINT(GREEN, " timeout(%zu sec) cgi %d -> kill, client: %p", current_time - timeout_limit, cgi_fd, client);
+        DEBUG_PRINT(RED, " timeout(%zu sec) cgi %d -> kill, client: %p", current_time - timeout_limit, cgi_fd, client);
         client->kill_cgi_process();
-        DEBUG_PRINT(GREEN, " cgi killed by signal read:%d, write:%d", client->cgi_read_fd(), client->cgi_write_fd());
+        DEBUG_PRINT(RED, " cgi killed by signal read:%d, write:%d", client->cgi_read_fd(), client->cgi_write_fd());
     }
+    // todo: erase cgi from timeout?
+}
 
-    // client event
-    // todo
+
+void Server::management_client_keepalive_timeout(time_t current_time) {
+    std::set<FdTimeoutLimitPair>::iterator client = this->keepalive_clients_.begin();
+    while (client != this->keepalive_clients_.end()) {
+        time_t timeout_limit = client->first;
+        DEBUG_SERVER_PRINT(" client_fd: %d, time limit: %zu, current: %zu -> %s",
+                           client->second, timeout_limit, current_time, (timeout_limit <= current_time ? "limited" : "ok"));
+        if (current_time < timeout_limit) {
+            DEBUG_PRINT(GREEN, " client %d: time remaining(%zu sec)", client->second, current_time - timeout_limit);
+            break;  // sorted
+        }
+
+        int client_fd = client->second;
+        std::map<Fd, Event *>::iterator timeout_event = this->client_events_.find(client_fd);
+        if (timeout_event == this->client_events_.end())  {
+            const std::string error_msg = CREATE_ERROR_INFO_STR("error: fd not found in client_events");
+            DEBUG_PRINT(RED, "%s", error_msg.c_str());
+            ++client;
+            continue;
+        }
+
+        std::set<FdTimeoutLimitPair>::iterator current = client;
+        ++client;
+
+        this->keepalive_clients_.erase(current);
+
+        delete_event(timeout_event);
+        DEBUG_PRINT(RED, " client %d: time remaining(%zu sec) -> deleted", client_fd, current_time - timeout_limit);
+    }
+}
+
+
+void Server::management_timeout_events() {
+    time_t current_time = std::time(NULL);
+    DEBUG_PRINT(GREEN, "management_timeout current: %zu", current_time);
+
+    management_cgi_executing_timeout(current_time);
+    management_client_keepalive_timeout(current_time);
 }
 
 
@@ -352,6 +394,9 @@ ServerResult Server::create_event(int socket_fd) {
     if (accept_result.is_err()) {
         const std::string error_msg = accept_result.err_value();
         return ServerResult::err(error_msg);
+    }
+    if (accept_result.ok_value() == ERR) {  // exceed max connection
+        return ServerResult::ok(OK);
     }
     int connect_fd = accept_result.ok_value();
     Result<int, std::string> non_block = Socket::set_fd_to_nonblock(connect_fd);
@@ -408,6 +453,31 @@ void Server::delete_event(std::map<Fd, Event *>::iterator event) {
     this->fds_->clear_fd(client_fd);
     delete client_event;
     this->client_events_.erase(event);
+
+    std::set<FdTimeoutLimitPair>::iterator itr = find_timeout_fd_pair(client_fd, this->keepalive_clients_);
+    if (itr != this->keepalive_clients_.end()) {
+        this->keepalive_clients_.erase(itr);
+    }
+}
+
+
+void Server::idling_event(Event *event) {
+    // select        -> polling; idoling event
+    // epoll, kqueus -> event driven; init event
+    if (!event) {
+        return;
+    }
+    int client_fd = event->client_fd();
+    update_fd_type(client_fd, kWriteFd, kReadFd);
+    event->set_event_phase(kReceivingRequest);
+    event->clear_request();
+    event->clear_response();
+
+    time_t timeout_limit = std::time(NULL) + this->config_.keepalive_timeout();
+    this->keepalive_clients_.insert(FdTimeoutLimitPair(timeout_limit, client_fd));
+
+    DEBUG_SERVER_PRINT("init event: client_fd %d, timeout: %zu", client_fd, timeout_limit);
+    DEBUG_SERVER_PRINT("------------------------------------------------------------------------------------------------");
 }
 
 
@@ -428,18 +498,6 @@ void Server::update_fd_type(int fd,
     } else if (update_to == kWriteFd) {
         this->fds_->register_write_fd(fd);
     }
-}
-
-
-void Server::init_event(Event *event) {
-    if (!event) {
-        return;
-    }
-    int client_fd = event->client_fd();
-    update_fd_type(client_fd, kWriteFd, kReadFd);
-    event->set_event_phase(kReceivingRequest);
-    event->clear_request();
-    event->clear_response();
 }
 
 
@@ -541,6 +599,11 @@ ServerResult Server::handle_client_event(int client_fd) {
         const std::string error_msg = event_result.err_value();
         return ServerResult::err(error_msg);
     }
+
+    if (is_idling_client(client_fd) && event_result.ok_value() != Idling) {
+        clear_from_keepalive_clients(client_fd);
+    }
+
     switch (event_result.ok_value()) {
         case Success: {
             break;
@@ -557,6 +620,9 @@ ServerResult Server::handle_client_event(int client_fd) {
             register_cgi_write_fd_to_event_manager(&client_event);
             // register_cgi_fds_to_event_manager(&client_event);
             this->fds_->clear_fd(client_fd);
+            return ServerResult::ok(OK);
+        }
+        case Idling: {
             return ServerResult::ok(OK);
         }
         case ConnectionClosed: {
@@ -581,8 +647,13 @@ ServerResult Server::handle_client_event(int client_fd) {
         case kEventCompleted: {
             std::ostringstream oss; oss << client_event;
             DEBUG_SERVER_PRINT("process_event(client) -> event completed: %s", oss.str().c_str());
-            delete_event(event);  // todo -> init_session & keep-alive
-            // init_session(client);
+            if (client_event->is_keepalive()) {
+                DEBUG_SERVER_PRINT(" -> idling event %zu time", this->config_.keepalive_timeout());
+                idling_event(client_event);
+            } else {
+                DEBUG_SERVER_PRINT(" -> delete event");
+                delete_event(event);
+            }
             break;
         }
         default:
@@ -655,6 +726,34 @@ ServerResult Server::handle_cgi_event(int cgi_fd) {
 }
 
 
+std::set<FdTimeoutLimitPair>::iterator Server::find_timeout_fd_pair(int fd, const std::set<FdTimeoutLimitPair> &pair) {
+    std::set<FdTimeoutLimitPair>::iterator itr;
+    for (itr = pair.begin(); itr != pair.end(); ++itr) {
+        if (itr->second == fd) {
+            return itr;
+        }
+    }
+    return pair.end();
+}
+
+
+bool Server::is_idling_client(int fd) {
+    std::set<FdTimeoutLimitPair>::iterator client = find_timeout_fd_pair(fd, this->keepalive_clients_);
+    return client != this->keepalive_clients_.end();
+}
+
+
+void Server::clear_from_keepalive_clients(int fd) {
+    std::set<FdTimeoutLimitPair>::iterator client = find_timeout_fd_pair(fd, this->keepalive_clients_);
+    if (client == this->keepalive_clients_.end()) {
+        return;
+    }
+
+    DEBUG_SERVER_PRINT("idling -> active: client_fd %d", fd);
+    this->keepalive_clients_.erase(client);
+}
+
+
 ServerResult Server::process_event(int ready_fd) {
     if (is_socket_fd(ready_fd)) {
         DEBUG_SERVER_PRINT("  ready_fd=socket ready_fd: %d -> create_event()", ready_fd);
@@ -678,7 +777,7 @@ ServerResult Server::accept_connect_fd(int socket_fd,
     (void)MAX_SESSION;
     // if (MAX_SESSION <= this->client_fds_.size()) {
     //     std::cerr << "[Server Error] exceed max connection" << std::endl;
-    //     return ServerResult::ok(OK);  // todo: continue, ok?
+    //     return ServerResult::ok(ERR);  // todo: continue, ok?
     // }
 
     SocketResult accept_result = Socket::accept(socket_fd, client_addr);
@@ -706,9 +805,18 @@ ServerResult Server::accept_connect_fd(int socket_fd,
 
 
 void Server::set_io_timeout() {
-    if (this->cgi_fds_.empty() && !this->echo_mode_on_) {
-        this->fds_->set_io_timeout(0);
-    } else {
-        this->fds_->set_io_timeout(100);
+    const int kEchoModeTimeoutMs = 100;
+    if (this->echo_mode_on_) {
+        this->fds_->set_io_timeout(kEchoModeTimeoutMs);
+        return;
     }
+
+    const int kManagemtntTimeoutMs = 500;
+    if (!this->cgi_fds_.empty() || !this->keepalive_clients_.empty()) {
+        this->fds_->set_io_timeout(kManagemtntTimeoutMs);
+        return;
+    }
+
+    const int kTimeoutInfinity = 0;
+    this->fds_->set_io_timeout(kTimeoutInfinity);
 }
